refactor(core): keep rectangle opacity policy in WebGL
Keep blending-state optimization with the only backend that consumes it.

Reuse existing WebGL rectangle shape checks instead of duplicating them in semantic marks.
