+++ packages/core/src/marks/rect.js
@@ -1,7 +1,6 @@
 import Mark from "./mark.js";
 import { fixCoveragePositional, fixFill, fixStroke } from "./markUtils.js";
-import { isNestedDiscreteOffsetDef, isValueDef } from "../encoder/encoder.js";
-import { getCachedOrCall } from "../utils/propertyCacher.js";
+import { isNestedDiscreteOffsetDef } from "../encoder/encoder.js";
 
 /** @extends {Mark<import("../spec/mark.js").RectProps>} */
 export default class RectMark extends Mark {
@@ -22,22 +21,6 @@ export default class RectMark extends Mark {
     /** @param {string} channel @returns {number} @protected */
     getOffsetBand(channel) {
         return channel == "x2Offset" || channel == "y2Offset" ? 1 : 0;
-    }
-
-    get opaque() {
-        return (
-            getCachedOrCall(
-                this,
-                "opaque",
-                () =>
-                    !this.#isRoundedCorners() &&
-                    !this.#isStroked() &&
-                    !this.properties.shadowOpacity &&
-                    isValueDef(this.encoding.fillOpacity) &&
-                    this.encoding.fillOpacity.value == 1.0 &&
-                    this.properties.minOpacity == 1.0
-            ) && this.unitView.getEffectiveOpacity() == 1
-        );
     }
 
     /**
@@ -61,24 +44,5 @@ export default class RectMark extends Mark {
         delete encoding.color;
         delete encoding.opacity;
         return encoding;
-    }
-
-    #isRoundedCorners() {
-        const p = this.properties;
-        return (
-            p.cornerRadius ||
-            p.cornerRadiusBottomLeft ||
-            p.cornerRadiusBottomRight ||
-            p.cornerRadiusTopLeft ||
-            p.cornerRadiusTopRight
-        );
-    }
-
-    #isStroked() {
-        const strokeWidth = this.encoding.strokeWidth;
-        return (
-            !(isValueDef(strokeWidth) && !strokeWidth.value) ||
-            "condition" in strokeWidth
-        );
     }
 }
